Split out DMA work
