+++ tests_and_snippets/c_and_cpp/dma/dma.hpp
@@ -1,0 +1,16 @@
+#ifndef _WSPR_H
+#define _WSPR_H
+
+#ifdef __cplusplus
+extern "C"
+{
+#include "mailbox.h"
+}
+#endif /* __cplusplus */
+
+void dma_cleanup();
+void setup_dma();
+void tx_tone();
+void tx_wspr();
+
+#endif // _WSPR_H
