+++ tests_and_snippets/c_and_cpp/dma/main.cpp
@@ -1,1422 +1,32 @@
 #include "main.hpp"
 
-// Note on accessing memory in RPi:
-//
-// There are 3 (yes three) address spaces in the Pi:
-// Physical addresses
-//   These are the actual address locations of the RAM and are equivalent
-//   to offsets into /dev/mem.
-//   The peripherals (DMA engine, PWM, etc.) are located at physical
-//   address 0x2000000 for RPi1 and 0x3F000000 for RPi2/3.
-// Virtual addresses
-//   These are the addresses that a program sees and can read/write to.
-//   Addresses 0x00000000 through 0xBFFFFFFF are the addresses available
-//   to a program running in user space.
-//   Addresses 0xC0000000 and above are available only to the kernel.
-//   The peripherals start at address 0xF2000000 in virtual space but
-//   this range is only accessible by the kernel. The kernel could directly
-//   access peripherals from virtual addresses. It is not clear to me my
-//   a user space application running as 'root' does not have access to this
-//   memory range.
-// Bus addresses
-//   This is a different (virtual?) address space that also maps onto
-//   physical memory.
-//   The peripherals start at address 0x7E000000 of the bus address space.
-//   The DRAM is also available in bus address space in 4 different locations:
-//   0x00000000 "L1 and L2 cached alias"
-//   0x40000000 "L2 cache coherent (non allocating)"
-//   0x80000000 "L2 cache (only)"
-//   0xC0000000 "Direct, uncached access"
-//
-// Accessing peripherals from user space (virtual addresses):
-//   The technique used in this program is that mmap is used to map portions of
-//   /dev/mem to an arbitrary virtual address. For example, to access the
-//   GPIO's, the gpio range of addresses in /dev/mem (physical addresses) are
-//   mapped to a kernel chosen virtual address. After the mapping has been
-//   set up, writing to the kernel chosen virtual address will actually
-//   write to the GPIO addresses in physical memory.
-//
-// Accessing RAM from DMA engine
-//   The DMA engine is programmed by accessing the peripheral registers but
-//   must use bus addresses to access memory. Thus, to use the DMA engine to
-//   move memory from one virtual address to another virtual address, one needs
-//   to first find the physical addresses that corresponds to the virtual
-//   addresses. Then, one needs to find the bus addresses that corresponds to
-//   those physical addresses. Finally, the DMA engine can be programmed. i.e.
-//   DMA engine access should use addresses starting with 0xC.
-//
-// The perhipherals in the Broadcom documentation are described using their bus
-// addresses and structures are created and calculations performed in this
-// program to figure out how to access them with virtual addresses.
-
-// Empirical value for F_PWM_CLK that produces WSPR symbols that are 'close' to
-// 0.682s long. For some reason, despite the use of DMA, the load on the PI
-// affects the TX length of the symbols. However, the varying symbol length is
-// compensated for in the main loop.
-#define F_PWM_CLK_INIT (31156186.6125761)
-
-// WSRP nominal symbol time
-#define WSPR_SYMTIME (8192.0 / 12000.0)
-// How much random frequency offset should be added to WSPR transmissions
-// if the --offset option has been turned on.
-#define WSPR_RAND_OFFSET 80
-#define WSPR15_RAND_OFFSET 8
-
-#define PAGE_SIZE (4 * 1024)
-#define BLOCK_SIZE (4 * 1024)
-
-#define PWM_CLOCKS_PER_ITER_NOMINAL 1000
-
-// Given an address in the bus address space of the peripherals, this
-// macro calculates the appropriate virtual address to use to access
-// the requested bus address space. It does this by first subtracting
-// 0x7e000000 from the supplied bus address to calculate the offset into
-// the peripheral address space. Then, this offset is added to peri_base_virt
-// Which is the base address of the peripherals, in virtual address space.
-#define ACCESS_BUS_ADDR(buss_addr) *(volatile int *)((long int)peri_base_virt + (buss_addr) - 0x7e000000)
-// Given a bus address in the peripheral address space, set or clear a bit.
-#define SETBIT_BUS_ADDR(base, bit) ACCESS_BUS_ADDR(base) |= 1 << bit
-#define CLRBIT_BUS_ADDR(base, bit) ACCESS_BUS_ADDR(base) &= ~(1 << bit)
-
-// The following are all bus addresses.
-#define GPIO_BUS_BASE (0x7E200000)
-#define CM_GP0CTL_BUS (0x7e101070)
-#define CM_GP0DIV_BUS (0x7e101074)
-#define PADS_GPIO_0_27_BUS (0x7e10002c)
-#define CLK_BUS_BASE (0x7E101000)
-#define DMA_BUS_BASE (0x7E007000)
-#define PWM_BUS_BASE (0x7e20C000) /* PWM controller */
-
-// Used for GPIO DIO Access:
-// GPIO setup macros. Always use INP_GPIO(x) before using OUT_GPIO(x) or SET_GPIO_ALT(x,y)
-#define INP_GPIO(g) *(gpio + ((g) / 10)) &= ~(7 << (((g) % 10) * 3))
-#define OUT_GPIO(g) *(gpio + ((g) / 10)) |= (1 << (((g) % 10) * 3))
-#define SET_GPIO_ALT(g, a) *(gpio + (((g) / 10))) |= (((a) <= 3 ? (a) + 4 : (a) == 4 ? 3  \
-                                                                                     : 2) \
-                                                      << (((g) % 10) * 3))
-#define GPIO_SET *(gpio + 7)                  // sets bits which are 1 ignores bits which are 0
-#define GPIO_CLR *(gpio + 10)                 // clears bits which are 1 ignores bits which are 0
-#define GET_GPIO(g) (*(gpio + 13) & (1 << g)) // 0 if LOW, (1<<g) if HIGH
-#define GPIO_PULL *(gpio + 37)                // Pull up/pull down
-#define GPIO_PULLCLK0 *(gpio + 38)            // Pull up/pull down clock
-
-// Convert from a bus address to a physical address.
-#define BUS_TO_PHYS(x) ((x) & ~0xC0000000)
-
-// peri_base_virt is the base virtual address that a userspace program (this
-// program) can use to read/write to the the physical addresses controlling
-// the peripherals. This address is mapped at runtime using mmap and /dev/mem.
-// This must be declared global so that it can be called by the atexit
-// function.
-volatile unsigned *peri_base_virt = NULL;
-
-// Used for GPIO DIO Access:
-// Map to GPIO register
-void *gpio_map;
-// I/O access
-volatile unsigned *gpio;
-
-typedef enum
-{
-    WSPR,
-    TONE
-} mode_type;
-
-struct GPCTL
-{
-    // Structure used to control clock generator
-    char SRC : 4;
-    char ENAB : 1;
-    char KILL : 1;
-    char : 1;
-    char BUSY : 1;
-    char FLIP : 1;
-    char MASH : 2;
-    unsigned int : 13;
-    char PASSWD : 8;
-};
-
-struct CB
-{
-    // Structure used to tell the DMA engine what to do
-    volatile unsigned int TI;
-    volatile unsigned int SOURCE_AD;
-    volatile unsigned int DEST_AD;
-    volatile unsigned int TXFR_LEN;
-    volatile unsigned int STRIDE;
-    volatile unsigned int NEXTCONBK;
-    volatile unsigned int RES1;
-    volatile unsigned int RES2;
-};
-
-struct DMAregs
-{
-    // DMA engine status registers
-    volatile unsigned int CS;
-    volatile unsigned int CONBLK_AD;
-    volatile unsigned int TI;
-    volatile unsigned int SOURCE_AD;
-    volatile unsigned int DEST_AD;
-    volatile unsigned int TXFR_LEN;
-    volatile unsigned int STRIDE;
-    volatile unsigned int NEXTCONBK;
-    volatile unsigned int DEBUG;
-};
-
-struct PageInfo
-{
-    // Virtual and bus addresses of a page of physical memory.
-    void *b; // bus address
-    void *v; // virtual address
-};
-
-static struct
-{
-    // Must be global so that exit handlers can access this.
-    int handle;                                          /* From mbox_open() */
-    unsigned mem_ref = 0;                                /* From mem_alloc() */
-    unsigned bus_addr;                                   /* From mem_lock() */
-    unsigned char *virt_addr = NULL; /* From mapmem() */ // ha7ilm: originally uint8_t
-    unsigned pool_size;
-    unsigned pool_cnt;
-} mbox;
-
-struct wConfig
-{
-    // Global configuration items from command line and ini file
-    bool useini = false;
-    std::string inifile = "";                  // Default to empty, meaning no INI file specified
-    bool xmit_enabled = true;                  // Transmission disabled by default
-    bool repeat = true;                        // No repeat transmission by default
-    std::string callsign = "AA0NT";            // Default to empty, requiring user input
-    std::string locator = "EM18";              // Default to empty, requiring user input
-    std::string tx_power = "20";               // Default to 37 dBm (5W), a common WSPR power level
-    std::string frequency_string = "7040100."; // Default to empty
-    std::vector<double> center_freq_set = {};  // Empty vector, frequencies to be defined
-    double ppm = 0.0;                          // Default to zero, meaning no frequency correction applied
-    bool self_cal = true;                      // Self-calibration enabled by default
-    bool random_offset = true;                 // No random offset by default
-    double test_tone = 7040100;                // Default to NAN, meaning no test tone
-    bool no_delay = false;                     // Delay enabled by default
-    mode_type mode = WSPR;                     // Default mode is WSPR
-    int terminate = -1;                        // -1 to indicate no termination signal
-    bool useled = false;                       // No LED signaling by default
-    bool daemon_mode = false;                  // Not running as a daemon by default
-    double f_plld_clk = 125e6;                 // Default PLLD clock frequency: 125 MHz
-    int mem_flag = 0;                          // Default memory flag set to 0
-} config;
-
-static unsigned get_dt_ranges(const char *filename, unsigned offset)
-{
-    unsigned address = ~0;
-    FILE *fp = fopen(filename, "rb");
-    if (fp)
-    {
-        unsigned char buf[4];
-        fseek(fp, offset, SEEK_SET);
-        if (fread(buf, 1, sizeof buf, fp) == sizeof buf)
-            address = buf[0] << 24 | buf[1] << 16 | buf[2] << 8 | buf[3] << 0;
-        fclose(fp);
-    }
-    return address;
-}
-
-unsigned bcm_host_get_peripheral_address(void)
-{
-    unsigned address = get_dt_ranges("/proc/device-tree/soc/ranges", 4);
-    if (address == 0)
-        address = get_dt_ranges("/proc/device-tree/soc/ranges", 8);
-
-    return address == (unsigned)~0 ? 0x20000000 : address;
-}
-
-unsigned gpioBase()
-{
-    return bcm_host_get_peripheral_address();
-}
-
-static int read_string_from_file(const char *filename, const char *format, unsigned int *value)
-{
-    FILE *fin;
-    char str[256];
-    int found = 0;
-
-    fin = fopen(filename, "rt");
-
-    if (fin == NULL)
-        return 0;
-
-    while (fgets(str, sizeof(str), fin) != NULL)
-    {
-        if (value)
-        {
-            if (sscanf(str, format, value) == 1)
-            {
-                found = 1;
-                break;
-            }
-        }
-        else
-        {
-            if (!strcmp(str, format))
-            {
-                found = 1;
-                break;
-            }
-        }
-    }
-
-    fclose(fin);
-
-    return found;
-}
-
-static unsigned int get_revision_code()
-{
-    static unsigned int revision_num = (unsigned int)-1; // Ensure proper unsigned initialization
-    unsigned int num;
-
-    if (revision_num == (unsigned int)-1 && read_string_from_file("/proc/cpuinfo", "Revision : %x", &num))
-        revision_num = num;
-
-    return revision_num;
-}
-
-/* returns the processor ID
- */
-#define BCM_HOST_PROCESSOR_BCM2835 0
-#define BCM_HOST_PROCESSOR_BCM2836 1
-#define BCM_HOST_PROCESSOR_BCM2837 2
-#define BCM_HOST_PROCESSOR_BCM2838 3 /* Deprecated name */
-#define BCM_HOST_PROCESSOR_BCM2711 3
-
-/* returns the processor ID
- */
-int bcm_host_get_processor_id(void)
-{
-    unsigned int revision_num = get_revision_code();
-
-    if (revision_num & 0x800000)
-    {
-        return (revision_num & 0xf000) >> 12;
-    }
-    else
-    {
-        // Old style number only used 2835
-        return BCM_HOST_PROCESSOR_BCM2835;
-    }
-}
-
-int ver()
-{
-    // Return version ID
-    // See: https://github.com/raspberrypi/firmware/blob/5325b7802ca90ac5c87440e6acbc37c96f08b054/opt/vc/include/bcm_host.h#L93-L99
-    return bcm_host_get_processor_id();
-}
-
-// Recursive variadic functions for stdout/err control
-template <typename T>
-void prtStdOut(T t)
-{
-    std::cout << t;
-}
-template <typename T, typename... Args>
-void prtStdOut(T t, Args... args)
-{
-    std::cout << t;
-    prtStdOut(args...);
-}
-
-template <typename T>
-void prtStdErr(T t)
-{
-    std::cerr << t;
-}
-template <typename T, typename... Args>
-void prtStdErr(T t, Args... args)
-{
-    std::cerr << t;
-    prtStdErr(args...);
-}
-
-// GPIO/DIO Control:
-
-void setupGPIO(int pin = 0)
-{
-    if (pin == 0)
-        return;
-    // Set up gpio pointer for direct register access
-    int mem_fd;
-    // Set up a memory regions to access GPIO
-    unsigned gpio_base = gpioBase() + 0x200000;
-
-    if ((mem_fd = open("/dev/mem", O_RDWR | O_SYNC)) < 0)
-    {
-        prtStdErr("Fail: Unable to open /dev/mem (running as root?)\n");
-        exit(-1);
-    }
-
-    /* mmap GPIO */
-    gpio_map = mmap(
-        NULL,                   // Any adddress in our space will do
-        BLOCK_SIZE,             // Map length
-        PROT_READ | PROT_WRITE, // Enable reading & writting to mapped memory
-        MAP_SHARED,             // Shared with other processes
-        mem_fd,                 // File to map
-        gpio_base               // Offset to GPIO peripheral
-    );
-
-    close(mem_fd); // No need to keep mem_fd open after mmap
-
-    if (gpio_map == MAP_FAILED)
-    {
-        printf("Fail: mmap error %d\n", (int)gpio_map); // errno also set
-        exit(-1);
-    }
-
-    // Always use volatile pointer
-    gpio = (volatile unsigned *)gpio_map;
-
-    // Set GPIO pins to output
-    // Must use INP_GPIO before we can use OUT_GPIO
-    INP_GPIO(pin);
-    OUT_GPIO(pin);
-}
-
-void pinHigh(int pin = 0)
-{
-    if (pin == 0)
-        return;
-    GPIO_SET = 1 << pin;
-}
-
-void pinLow(int pin = 0)
-{
-    if (pin == 0)
-        return;
-    GPIO_CLR = 1 << pin;
-}
-
-// GPIO/DIO Control^
-
-void getPLLD()
-{
-    // Nominal clock frequencies
-    // double f_xtal = 19200000.0;
-    // PLLD clock frequency.
-    // For RPi1, after NTP converges, these is a 2.5 PPM difference between
-    // the PPM correction reported by NTP and the actual frequency offset of
-    // the crystal. This 2.5 PPM offset is not present in the RPi2 and RPi3 (RPI4).
-    // This 2.5 PPM offset is compensated for here, but only for the RPi1.
-
-    switch (ver())
-    {
-    case 0: // RPi1
-        config.mem_flag = 0x0c;
-        config.f_plld_clk = (500000000.0 * (1 - 2.500e-6));
-        break;
-    case 1: // RPi2
-    case 2: // RPi3
-        config.mem_flag = 0x04;
-        config.f_plld_clk = (500000000.0);
-        break;
-    case 3: // RPi 4
-        config.mem_flag = 0x04;
-        config.f_plld_clk = (750000000.0);
-        break;
-    default:
-        fprintf(stderr, "Error: Unknown chipset (%d).", ver());
-        exit(-1);
-    }
-    if (config.f_plld_clk <= 0)
-    {
-        prtStdErr("Error: Invalid PLL clock frequency. Using default 500MHz.\n");
-        config.f_plld_clk = 500000000.0;
-    }
-}
-
-void allocMemPool(unsigned numpages)
-{
-    // Use the mbox interface to allocate a single chunk of memory to hold
-    // all the pages we will need. The bus address and the virtual address
-    // are saved in the mbox structure.
-
-    // Allocate space.
-    mbox.mem_ref = mem_alloc(mbox.handle, 4096 * numpages, 4096, config.mem_flag);
-    // Lock down the allocated space and return its bus address.
-    mbox.bus_addr = mem_lock(mbox.handle, mbox.mem_ref);
-    // Conert the bus address to a physical address and map this to virtual
-    // (aka user) space.
-    mbox.virt_addr = (unsigned char *)mapmem(BUS_TO_PHYS(mbox.bus_addr), 4096 * numpages);
-    // The number of pages in the pool. Never changes
-    mbox.pool_size = numpages;
-    // How many of the created pages have actually been used.
-    mbox.pool_cnt = 0;
-    // printf("allocMemoryPool bus_addr=%x virt_addr=%x mem_ref=%x\n",mbox.bus_addr,(unsigned)mbox.virt_addr,mbox.mem_ref);
-}
-
-void getRealMemPageFromPool(void **vAddr, void **bAddr)
-{
-    // Returns the virtual and bus address (NOT physical address) of another
-    // page in the pool.
-    if (mbox.pool_cnt >= mbox.pool_size)
-    {
-        prtStdErr("Error: unable to allocated more pages.");
-        exit(-1);
-    }
-    unsigned offset = mbox.pool_cnt * 4096;
-    *vAddr = (void *)(((unsigned)mbox.virt_addr) + offset);
-    *bAddr = (void *)(((unsigned)mbox.bus_addr) + offset);
-    // printf("getRealMemoryPageFromPool bus_addr=%x virt_addr=%x\n", (unsigned)*pAddr,(unsigned)*vAddr);
-    mbox.pool_cnt++;
-}
-
-void deallocMemPool()
-{
-    // Free the memory pool
-    if (mbox.virt_addr != NULL)
-    {
-        unmapmem(mbox.virt_addr, mbox.pool_size * 4096);
-    }
-    if (mbox.mem_ref != 0)
-    {
-        mem_unlock(mbox.handle, mbox.mem_ref);
-        mem_free(mbox.handle, mbox.mem_ref);
-    }
-}
-
-void disable_clock()
-{
-    // Disable the PWM clock and wait for it to become 'not busy'.
-    // Check if mapping has been set up yet.
-    if (peri_base_virt == NULL)
-    {
-        return;
-    }
-    // Disable the clock (in case it's already running) by reading current
-    // settings and only clearing the enable bit.
-    auto settings = ACCESS_BUS_ADDR(CM_GP0CTL_BUS);
-    // Clear enable bit and add password
-    settings = (settings & 0x7EF) | 0x5A000000;
-    // Disable
-    ACCESS_BUS_ADDR(CM_GP0CTL_BUS) = *((int *)&settings);
-    // Wait for clock to not be busy.
-    while (true)
-    {
-        if (!(ACCESS_BUS_ADDR(CM_GP0CTL_BUS) & (1 << 7)))
-        {
-            break;
-        }
-    }
-}
-
-void txon(bool led = false)
-{
-    // Turn on TX
-    // Set function select for GPIO4.
-    // Fsel 000 => input
-    // Fsel 001 => output
-    // Fsel 100 => alternate function 0
-    // Fsel 101 => alternate function 1
-    // Fsel 110 => alternate function 2
-    // Fsel 111 => alternate function 3
-    // Fsel 011 => alternate function 4
-    // Fsel 010 => alternate function 5
-    // Function select for GPIO is configured as 'b100 which selects
-    // alternate function 0 for GPIO4. Alternate function 0 is GPCLK0.
-    // See section 6.2 of Arm Peripherals Manual.
-    SETBIT_BUS_ADDR(GPIO_BUS_BASE, 14);
-    CLRBIT_BUS_ADDR(GPIO_BUS_BASE, 13);
-    CLRBIT_BUS_ADDR(GPIO_BUS_BASE, 12);
-
-    // Set GPIO drive strength, more info: http://www.scribd.com/doc/101830961/GPIO-Pads-Control2
-    // ACCESS_BUS_ADDR(PADS_GPIO_0_27_BUS) = 0x5a000018 + 0;  //2mA -3.4dBm
-    // ACCESS_BUS_ADDR(PADS_GPIO_0_27_BUS) = 0x5a000018 + 1;  //4mA +2.1dBm
-    // ACCESS_BUS_ADDR(PADS_GPIO_0_27_BUS) = 0x5a000018 + 2;  //6mA +4.9dBm
-    // ACCESS_BUS_ADDR(PADS_GPIO_0_27_BUS) = 0x5a000018 + 3;  //8mA +6.6dBm(default)
-    // ACCESS_BUS_ADDR(PADS_GPIO_0_27_BUS) = 0x5a000018 + 4;  //10mA +8.2dBm
-    // ACCESS_BUS_ADDR(PADS_GPIO_0_27_BUS) = 0x5a000018 + 5;  //12mA +9.2dBm
-    // ACCESS_BUS_ADDR(PADS_GPIO_0_27_BUS) = 0x5a000018 + 6;  //14mA +10.0dBm
-    ACCESS_BUS_ADDR(PADS_GPIO_0_27_BUS) = 0x5a000018 + 7; // 16mA +10.6dBm
-
-    disable_clock();
-
-    // Set clock source as PLLD.
-    struct GPCTL setupword = {6 /*SRC*/, 0, 0, 0, 0, 3, 0x5a};
-
-    // Enable clock.
-    setupword = {6 /*SRC*/, 1, 0, 0, 0, 3, 0x5a};
-    int temp;
-    memcpy(&temp, &setupword, sizeof(int));
-    ACCESS_BUS_ADDR(CM_GP0CTL_BUS) = temp;
-}
-
-void txoff(bool led = false)
-{
-    // Turn transmitter on
-    // struct GPCTL setupword = {6/*SRC*/, 0, 0, 0, 0, 1,0x5a};
-    // ACCESS_BUS_ADDR(CM_GP0CTL_BUS) = *((int*)&setupword);
-    disable_clock();
-}
-
-void txSym(
-    const int &sym_num,
-    const double &center_freq,
-    const double &tone_spacing,
-    const double &tsym,
-    const std::vector<double> &dma_table_freq,
-    const double &f_pwm_clk,
-    struct PageInfo instrs[],
-    struct PageInfo &constPage,
-    int &bufPtr)
-{
-    // Transmit symbol sym for tsym seconds.
-    //
-    // Note:
-    // Upon entering this function at the beginning of a WSPR transmission, we
-    // do not know which DMA table entry is being processed by the DMA engine.
-    const int f0_idx = sym_num * 2;
-    const int f1_idx = f0_idx + 1;
-    const double f0_freq = dma_table_freq[f0_idx];
-    const double f1_freq = dma_table_freq[f1_idx];
-    const double tone_freq = center_freq - 1.5 * tone_spacing + sym_num * tone_spacing;
-    // Double check.
-    assert((tone_freq >= f0_freq) && (tone_freq <= f1_freq));
-    const double f0_ratio = 1.0 - (tone_freq - f0_freq) / (f1_freq - f0_freq);
-    // cout << "f0_ratio = " << f0_ratio << "\n";
-    assert((f0_ratio >= 0) && (f0_ratio <= 1));
-    const long int n_pwmclk_per_sym = round(f_pwm_clk * tsym);
-
-    long int n_pwmclk_transmitted = 0;
-    long int n_f0_transmitted = 0;
-    // printf("<instrs[bufPtr] begin=%x>",(unsigned)&instrs[bufPtr]);
-    while (n_pwmclk_transmitted < n_pwmclk_per_sym)
-    {
-        // Number of PWM clocks for this iteration
-        long int n_pwmclk = PWM_CLOCKS_PER_ITER_NOMINAL;
-        // Iterations may produce spurs around the main peak based on the iteration
-        // frequency. Randomize the iteration period so as to spread this peak
-        // around.
-        n_pwmclk += round((rand() / ((double)RAND_MAX + 1.0) - .5) * n_pwmclk) * 1;
-        if (n_pwmclk_transmitted + n_pwmclk > n_pwmclk_per_sym)
-        {
-            n_pwmclk = n_pwmclk_per_sym - n_pwmclk_transmitted;
-        }
-
-        // Calculate number of clocks to transmit f0 during this iteration so
-        // that the long term average is as close to f0_ratio as possible.
-        const long int n_f0 = round(f0_ratio * (n_pwmclk_transmitted + n_pwmclk)) - n_f0_transmitted;
-        const long int n_f1 = n_pwmclk - n_f0;
-
-        // Configure the transmission for this iteration
-        // Set GPIO pin to transmit f0
-        bufPtr++;
-        while (ACCESS_BUS_ADDR(DMA_BUS_BASE + 0x04 /* CurBlock*/) == (long int)(instrs[bufPtr].b))
-            usleep(100);
-        ((struct CB *)(instrs[bufPtr].v))->SOURCE_AD = (long int)constPage.b + f0_idx * 4;
-
-        // Wait for n_f0 PWM clocks
-        bufPtr++;
-        while (ACCESS_BUS_ADDR(DMA_BUS_BASE + 0x04 /* CurBlock*/) == (long int)(instrs[bufPtr].b))
-            usleep(100);
-        ((struct CB *)(instrs[bufPtr].v))->TXFR_LEN = n_f0;
-
-        // Set GPIO pin to transmit f1
-        bufPtr++;
-        while (ACCESS_BUS_ADDR(DMA_BUS_BASE + 0x04 /* CurBlock*/) == (long int)(instrs[bufPtr].b))
-            usleep(100);
-        ((struct CB *)(instrs[bufPtr].v))->SOURCE_AD = (long int)constPage.b + f1_idx * 4;
-
-        // Wait for n_f1 PWM clocks
-        bufPtr = (bufPtr + 1) % (1024);
-        while (ACCESS_BUS_ADDR(DMA_BUS_BASE + 0x04 /* CurBlock*/) == (long int)(instrs[bufPtr].b))
-            usleep(100);
-        ((struct CB *)(instrs[bufPtr].v))->TXFR_LEN = n_f1;
-
-        // Update counters
-        n_pwmclk_transmitted += n_pwmclk;
-        n_f0_transmitted += n_f0;
-    }
-    // printf("<instrs[bufPtr]=%x %x>",(unsigned)instrs[bufPtr].v,(unsigned)instrs[bufPtr].b);
-}
-
-void unSetupDMA()
-{
-    // Turn off (reset) DMA engine
-    // Check if mapping has been set up yet.
-    if (peri_base_virt == NULL)
-    {
-        return;
-    }
-    // cout << "Exiting." << "\n";
-    struct DMAregs *DMA0 = (struct DMAregs *)&(ACCESS_BUS_ADDR(DMA_BUS_BASE));
-    DMA0->CS = 1 << 31; // reset dma controller
-    txoff();
-}
-
-double bit_trunc(const double &d, const int &lsb)
-{
-    // Truncate at bit lsb. i.e. set all bits less than lsb to zero.
-    return floor(d / pow(2.0, lsb)) * pow(2.0, lsb);
-}
-
-void setupDMATab(
-    const double &center_freq_desired,
-    const double &tone_spacing,
-    const double &plld_actual_freq,
-    std::vector<double> &dma_table_freq,
-    double &center_freq_actual,
-    struct PageInfo &constPage)
-{
-    if (dma_table_freq.size() < 1024)
-    {
-        prtStdErr("Error: DMA table frequency array not initialized properly.\n");
-        dma_table_freq.resize(1024);
-    }
-    // Program the tuning words into the DMA table.
-
-    // Make sure that all the WSPR tones can be produced solely by
-    // varying the fractional part of the frequency divider.
-    center_freq_actual = center_freq_desired;
-    double div_lo = bit_trunc(plld_actual_freq / (center_freq_desired - 1.5 * tone_spacing), -12) + pow(2.0, -12);
-    double div_hi = bit_trunc(plld_actual_freq / (center_freq_desired + 1.5 * tone_spacing), -12);
-    if (floor(div_lo) != floor(div_hi))
-    {
-        center_freq_actual = plld_actual_freq / floor(div_lo) - 1.6 * tone_spacing;
-        std::stringstream temp;
-        temp << std::setprecision(6) << std::fixed << "Warning: center frequency has been changed to " << center_freq_actual / 1e6 << " MHz";
-        prtStdErr(temp.str(), " because of hardware limitations.\n");
-    }
-
-    // Create DMA table of tuning words. WSPR tone i will use entries 2*i and
-    // 2*i+1 to generate the appropriate tone.
-    double tone0_freq = center_freq_actual - 1.5 * tone_spacing;
-    std::vector<long int> tuning_word(1024);
-    for (int i = 0; i < 8; i++)
-    {
-        double tone_freq = tone0_freq + (i >> 1) * tone_spacing;
-        double div = bit_trunc(plld_actual_freq / tone_freq, -12);
-        if (i % 2 == 0)
-        {
-            div = div + pow(2.0, -12);
-        }
-        tuning_word[i] = ((int)(div * pow(2.0, 12)));
-    }
-    // Fill the remaining table, just in case.
-    for (int i = 8; i < 1024; i++)
-    {
-        double div = 500 + i;
-        tuning_word[i] = ((int)(div * pow(2.0, 12)));
-    }
-
-    // Program the table
-    dma_table_freq.resize(1024);
-    for (int i = 0; i < 1024; i++)
-    {
-        dma_table_freq[i] = plld_actual_freq / (tuning_word[i] / pow(2.0, 12));
-        ((int *)(constPage.v))[i] = (0x5a << 24) + tuning_word[i];
-        if ((i % 2 == 0) && (i < 8))
-        {
-            assert((tuning_word[i] & (~0xfff)) == (tuning_word[i + 1] & (~0xfff)));
-        }
-    }
-}
-
-void setupDMA(
-    struct PageInfo &constPage,
-    struct PageInfo &instrPage,
-    struct PageInfo instrs[])
-{
-    // Create the memory structures needed by the DMA engine and perform initial
-    // clock configuration.
-
-    allocMemPool(1025);
-
-    // Allocate a page of ram for the constants
-    getRealMemPageFromPool(&constPage.v, &constPage.b);
-
-    // Create 1024 instructions allocating one page at a time.
-    // Even instructions target the GP0 Clock divider
-    // Odd instructions target the PWM FIFO
-    int instrCnt = 0;
-    while (instrCnt < 1024)
-    {
-        // Allocate a page of ram for the instructions
-        getRealMemPageFromPool(&instrPage.v, &instrPage.b);
-
-        // Make copy instructions
-        // Only create as many instructions as will fit in the recently
-        // allocated page. If not enough space for all instructions, the
-        // next loop will allocate another page.
-        struct CB *instr0 = (struct CB *)instrPage.v;
-        int i;
-        for (i = 0; i < (signed)(4096 / sizeof(struct CB)); i++)
-        {
-            instrs[instrCnt].v = (void *)((long int)instrPage.v + sizeof(struct CB) * i);
-            instrs[instrCnt].b = (void *)((long int)instrPage.b + sizeof(struct CB) * i);
-            instr0->SOURCE_AD = (unsigned long int)constPage.b + 2048;
-            instr0->DEST_AD = PWM_BUS_BASE + 0x18 /* FIF1 */;
-            instr0->TXFR_LEN = 4;
-            instr0->STRIDE = 0;
-            // instr0->NEXTCONBK = (int)instrPage.b + sizeof(struct CB)*(i+1);
-            instr0->TI = (1 /* DREQ  */ << 6) | (5 /* PWM */ << 16) | (1 << 26 /* no wide*/);
-            instr0->RES1 = 0;
-            instr0->RES2 = 0;
-
-            // Shouldn't this be (instrCnt%2) ???
-            if (i % 2)
-            {
-                instr0->DEST_AD = CM_GP0DIV_BUS;
-                instr0->STRIDE = 4;
-                instr0->TI = (1 << 26 /* no wide*/);
-            }
-
-            if (instrCnt != 0)
-                ((struct CB *)(instrs[instrCnt - 1].v))->NEXTCONBK = (long int)instrs[instrCnt].b;
-            instr0++;
-            instrCnt++;
-        }
-    }
-    // Create a circular linked list of instructions
-    ((struct CB *)(instrs[1023].v))->NEXTCONBK = (long int)instrs[0].b;
-
-    // set up a clock for the PWM
-    ACCESS_BUS_ADDR(CLK_BUS_BASE + 40 * 4 /*PWMCLK_CNTL*/) = 0x5A000026; // Source=PLLD and disable
-    usleep(1000);
-    // ACCESS_BUS_ADDR(CLK_BUS_BASE + 41*4 /*PWMCLK_DIV*/)  = 0x5A002800;
-    ACCESS_BUS_ADDR(CLK_BUS_BASE + 41 * 4 /*PWMCLK_DIV*/) = 0x5A002000;  // set PWM div to 2, for 250MHz
-    ACCESS_BUS_ADDR(CLK_BUS_BASE + 40 * 4 /*PWMCLK_CNTL*/) = 0x5A000016; // Source=PLLD and enable
-    usleep(1000);
-
-    // set up pwm
-    ACCESS_BUS_ADDR(PWM_BUS_BASE + 0x0 /* CTRL*/) = 0;
-    usleep(1000);
-    ACCESS_BUS_ADDR(PWM_BUS_BASE + 0x4 /* status*/) = -1; // clear errors
-    usleep(1000);
-    // Range should default to 32, but it is set at 2048 after reset on my RPi.
-    ACCESS_BUS_ADDR(PWM_BUS_BASE + 0x10) = 32;
-    ACCESS_BUS_ADDR(PWM_BUS_BASE + 0x20) = 32;
-    ACCESS_BUS_ADDR(PWM_BUS_BASE + 0x0 /* CTRL*/) = -1; //(1<<13 /* Use fifo */) | (1<<10 /* repeat */) | (1<<9 /* serializer */) | (1<<8 /* enable ch */) ;
-    usleep(1000);
-    ACCESS_BUS_ADDR(PWM_BUS_BASE + 0x8 /* DMAC*/) = (1 << 31 /* DMA enable */) | 0x0707;
-
-    // activate dma
-    struct DMAregs *DMA0 = (struct DMAregs *)&(ACCESS_BUS_ADDR(DMA_BUS_BASE));
-    DMA0->CS = 1 << 31; // reset
-    DMA0->CONBLK_AD = 0;
-    DMA0->TI = 0;
-    DMA0->CONBLK_AD = (unsigned long int)(instrPage.b);
-    DMA0->CS = (1 << 0) | (255 << 16); // enable bit = 0, clear end flag = 1, prio=19-16
-}
-
-void to_upper(char *str)
-{
-    // Convert string to uppercase
-    while (*str)
-    {
-        *str = toupper(*str);
-        str++;
-    }
-}
-
-void wspr(
-    const char *call,
-    const char *l_pre,
-    const char *dbm,
-    unsigned char *symbols)
-{
-    // Encode call, locator, and dBm into WSPR codeblock.
-
-    // pack prefix in nadd, call in n1, grid, dbm in n2
-    char *c, buf[16];
-    strncpy(buf, call, sizeof(buf) - 1);
-    buf[sizeof(buf) - 1] = '\0'; // Ensure null-termination
-    c = buf;
-    to_upper(c);
-    unsigned long ng, nadd = 0;
-
-    if (strchr(c, '/'))
-    { // prefix-suffix
-        nadd = 2;
-        int i = strchr(c, '/') - c; // stroke position
-        int n = strlen(c) - i - 1;  // suffix len, prefix-call len
-        c[i] = '\0';
-        if (n == 1)
-            ng = 60000 - 32768 + (c[i + 1] >= '0' && c[i + 1] <= '9' ? c[i + 1] - '0' : c[i + 1] == ' ' ? 38
-                                                                                                        : c[i + 1] - 'A' + 10); // suffix /A to /Z, /0 to /9
-        if (n == 2)
-            ng = 60000 + 26 + 10 * (c[i + 1] - '0') + (c[i + 2] - '0'); // suffix /10 to /99
-        if (n > 2)
-        { // prefix EA8/, right align
-            ng = (i < 3 ? 36 : c[i - 3] >= '0' && c[i - 3] <= '9' ? c[i - 3] - '0'
-                                                                  : c[i - 3] - 'A' + 10);
-            ng = 37 * ng + (i < 2 ? 36 : c[i - 2] >= '0' && c[i - 2] <= '9' ? c[i - 2] - '0'
-                                                                            : c[i - 2] - 'A' + 10);
-            ng = 37 * ng + (i < 1 ? 36 : c[i - 1] >= '0' && c[i - 1] <= '9' ? c[i - 1] - '0'
-                                                                            : c[i - 1] - 'A' + 10);
-            if (ng < 32768)
-                nadd = 1;
-            else
-                ng = ng - 32768;
-            c = c + i + 1;
-        }
-    }
-
-    int i = (isdigit(c[2]) ? 2 : isdigit(c[1]) ? 1
-                                               : 0); // last prefix digit of de-suffixed/de-prefixed callsign
-    int n = strlen(c) - i - 1;                       // 2nd part of call len
-    unsigned long n1;
-    n1 = (i < 2 ? 36 : c[i - 2] >= '0' && c[i - 2] <= '9' ? c[i - 2] - '0'
-                                                          : c[i - 2] - 'A' + 10);
-    n1 = 36 * n1 + (i < 1 ? 36 : c[i - 1] >= '0' && c[i - 1] <= '9' ? c[i - 1] - '0'
-                                                                    : c[i - 1] - 'A' + 10);
-    n1 = 10 * n1 + c[i] - '0';
-    n1 = 27 * n1 + (n < 1 ? 26 : c[i + 1] - 'A');
-    n1 = 27 * n1 + (n < 2 ? 26 : c[i + 2] - 'A');
-    n1 = 27 * n1 + (n < 3 ? 26 : c[i + 3] - 'A');
-
-    // if(rand() % 2) nadd=0;
-    if (!nadd)
-    {
-        // Copy locator locally since it is declared const and we cannot modify
-        // its contents in-place.
-        char l[5]; // Increase size to hold null terminator
-        strncpy(l, l_pre, sizeof(l) - 1);
-        l[sizeof(l) - 1] = '\0'; // Ensure null-termination
-        to_upper(l);             // grid square Maidenhead locator (uppercase)
-        ng = 180 * (179 - 10 * (l[0] - 'A') - (l[2] - '0')) + 10 * (l[1] - 'A') + (l[3] - '0');
-    }
-    int p = atoi(dbm); // EIRP in dBm={0,3,7,10,13,17,20,23,27,30,33,37,40,43,47,50,53,57,60}
-    int corr[] = {0, -1, 1, 0, -1, 2, 1, 0, -1, 1};
-    p = p > 60 ? 60 : p < 0 ? 0
-                            : p + corr[p % 10];
-    unsigned long n2 = (ng << 7) | (p + 64 + nadd);
-
-    // pack n1,n2,zero-tail into 50 bits
-    char packed[11] = {
-        static_cast<char>(n1 >> 20),
-        static_cast<char>(n1 >> 12),
-        static_cast<char>(n1 >> 4),
-        static_cast<char>(((n1 & 0x0f) << 4) | ((n2 >> 18) & 0x0f)),
-        static_cast<char>(n2 >> 10),
-        static_cast<char>(n2 >> 2),
-        static_cast<char>((n2 & 0x03) << 6),
-        0,
-        0,
-        0,
-        0};
-
-    // convolutional encoding K=32, r=1/2, Layland-Lushbaugh polynomials
-    int k = 0;
-    int j, s;
-    int nstate = 0;
-    unsigned char symbol[176];
-    for (j = 0; j != sizeof(packed); j++)
-    {
-        for (i = 7; i >= 0; i--)
-        {
-            unsigned long poly[2] = {0xf2d05351L, 0xe4613c47L};
-            nstate = (nstate << 1) | ((packed[j] >> i) & 1);
-            for (s = 0; s != 2; s++)
-            { // convolve
-                unsigned long n = nstate & poly[s];
-                int even = 0; // even := parity(n)
-                while (n)
-                {
-                    even = 1 - even;
-                    n = n & (n - 1);
-                }
-                symbol[k] = even;
-                k++;
-            }
-        }
-    }
-
-    // interleave symbols
-    const unsigned char npr3[162] = {
-        1, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0,
-        0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 0, 1, 0,
-        0, 0, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0, 1, 0, 1, 0,
-        0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 0, 1, 1, 1,
-        0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 0,
-        0, 0};
-    for (i = 0; i != 162; i++)
-    {
-        // j0 := bit reversed_values_smaller_than_161[i]
-        unsigned char j0 = 0; // Ensure j0 is initialized
-        p = -1;
-        for (k = 0; p != i; k++)
-        {
-            j0 = 0;                  // Reset j0 before each use
-            for (j = 0; j != 8; j++) // j0 := bit_reverse(k)
-                j0 = ((k >> j) & 1) | (j0 << 1);
-            if (j0 < 162)
-                p++;
-        }
-        symbols[j0] = npr3[j0] | (symbol[i] << 1); // interleave and add sync std::vector
-    }
-}
-
-void print_usage()
-{
-    prtStdOut("Usage:\n");
-    prtStdOut("  wspr [options] callsign locator tx_pwr_dBm f1 <f2> <f3> ...\n");
-    prtStdOut("    OR\n");
-    prtStdOut("  wspr [options] --test-tone f\n");
-    prtStdOut("\n");
-    prtStdOut("Options:\n");
-    prtStdOut("  -h --help\n");
-    prtStdOut("    Print out this help screen.\n");
-    prtStdOut("  -v --version\n");
-    prtStdOut("    Show Wsprry Pi version.\n");
-    prtStdOut("  -p --ppm ppm\n");
-    prtStdOut("    Known PPM correction to 19.2MHz RPi nominal crystal frequency.\n");
-    prtStdOut("  -s --self-calibration\n");
-    prtStdOut("    Check NTP before every transmission to obtain the PPM error of the\n");
-    prtStdOut("    crystal (default setting.).\n");
-    prtStdOut("  -f --free-running\n");
-    prtStdOut("    Do not use NTP to correct frequency error of RPi crystal.\n");
-    prtStdOut("  -r --repeat\n");
-    prtStdOut("    Repeatedly, and in order, transmit on all the specified command line\n");
-    prtStdOut("    freqs.\n");
-    prtStdOut("  -x --terminate <n>\n");
-    prtStdOut("    Terminate after n transmissions have been completed.\n");
-    prtStdOut("  -o --offset\n");
-    prtStdOut("    Add a random frequency offset to each transmission:\n");
-    prtStdOut("      +/- ", WSPR_RAND_OFFSET, " Hz for WSPR\n");
-    prtStdOut("      +/- ", WSPR15_RAND_OFFSET, " Hz for WSPR-15\n");
-    prtStdOut("  -t --test-tone freq\n");
-    prtStdOut("    Simply output a test tone at the specified frequency. Only used for\n");
-    prtStdOut("    debugging and to verify calibration.\n");
-    prtStdOut("  -l --led\n");
-    prtStdOut("    Use LED when transmitting (TAPR board).\n");
-    prtStdOut("  -n --no-delay\n");
-    prtStdOut("    Transmit immediately, do not wait for a WSPR TX window. Used for\n");
-    prtStdOut("    testing only.\n");
-    prtStdOut("  -i --ini-file\n");
-    prtStdOut("    Load parameters from an ini file. Supply path and file name.\n");
-    prtStdOut("  -D --daemon-mode\n");
-    prtStdOut("    Run with terse messaging.\n");
-    prtStdOut("\n");
-    prtStdOut("Frequencies can be specified either as an absolute TX carrier frequency,\n");
-    prtStdOut("or using one of the following strings:\n");
-    prtStdOut("\n");
-    prtStdOut("  LF, LF-15, MF, MF-15, 160m, 160m-15, 80m, 60m, 40m, 30m, 20m,\n");
-    prtStdOut("  17m, 15m, 12m, 10m, 6m, 4m, and 2m\n");
-    prtStdOut("\n");
-    prtStdOut("If a string is used, the transmission will happen in the middle of the\n");
-    prtStdOut("WSPR region of the selected band.\n");
-    prtStdOut("\n");
-    prtStdOut("The \"-15\" suffix indicates the WSPR-15 region of band.\n");
-    prtStdOut("\n");
-    prtStdOut("Transmission gaps can be created by specifying a TX frequency of 0.\n");
-    prtStdOut("\n");
-}
-
-bool wait_every(int minute)
-{
-    // Wait for the system clock's minute to reach one second past 'minute'
-    time_t t;
-    struct tm *ptm;
-    for (;;)
-    {
-        time(&t);
-        ptm = gmtime(&t);
-        if ((ptm->tm_min % minute) == 0 && ptm->tm_sec == 0)
-            break;
-        usleep(1000);
-    }
-    usleep(1000000); // Wait another second
-    return true;     // OK to proceed
-}
-
-void update_ppm()
-{
-    struct timex ntx;
-    int status;
-    double ppm_new;
-    int retry_count = 0;
-
-    while (retry_count < 5) {
-        ntx.modes = 0; /* only read */
-        status = ntp_adjtime(&ntx);
-
-        // Check if ntx.freq is within valid range
-        if (ntx.freq >= -500000 && ntx.freq <= 500000) {
-            break;  // Valid value, exit retry loop
-        }
-
-        prtStdErr("Warning: Invalid ntx.freq. Retrying in 2 seconds.\n");
-        sleep(2);
-        retry_count++;
-    }
-
-    // Final check before using ntx.freq
-    if (ntx.freq < -500000 || ntx.freq > 500000) {
-        prtStdErr("Error: ntx.freq remains out of range after retries. Using clamped value.\n");
-        ntx.freq = std::max(-500000L, std::min(ntx.freq, 500000L));
-    }
-
-    ppm_new = (double)ntx.freq / (double)(1 << 16);
-    
-    if (abs(ppm_new) > 200) {
-        prtStdErr("Warning: Absolute ppm value is greater than 200 and is being ignored.\n");
-    } else {
-        if (config.ppm != ppm_new) {
-            prtStdOut("Obtained new ppm value: ", ppm_new, "\n");
-        }
-        if (!std::isfinite(ppm_new) || ppm_new < -500 || ppm_new > 500) {
-            prtStdErr("Warning: Invalid PPM value. Clamping to ±500.\n");
-            config.ppm = std::max(-500.0, std::min(ppm_new, 500.0));
-        } else {
-            config.ppm = ppm_new;
-        }
-    }
-}
-
-int timeval_subtract(struct timeval *result, struct timeval *t2, struct timeval *t1)
-{
-    // Return 1 if the difference is negative, otherwise 0.
-    long int diff = (t2->tv_usec + 1000000 * t2->tv_sec) - (t1->tv_usec + 1000000 * t1->tv_sec);
-    result->tv_sec = diff / 1000000;
-    result->tv_usec = diff % 1000000;
-
-    return (diff < 0);
-}
-
-void timeval_print(struct timeval *tv)
-{
-    // Print Time Stamp
-
-    char buffer[30];
-    time_t curtime;
-
-    // printf("%ld.%06ld", tv->tv_sec, tv->tv_usec);
-    curtime = tv->tv_sec;
-    // strftime(buffer, 30, "%m-%d-%Y %T", localtime(&curtime));
-    strftime(buffer, 30, "%Y-%m-%d %T", gmtime(&curtime));
-    printf("%s.%03ld", buffer, (tv->tv_usec + 500) / 1000);
-    std::cout << " UTC";
-}
-
-void open_mbox()
-{
-    // Create the mbox special files and open mbox.
-    mbox.handle = mbox_open();
-    if (mbox.handle < 0)
-    {
-        prtStdErr("Failed to open mailbox.\n");
-        exit(-1);
-    }
-}
-
-void cleanup()
+#include <csignal>
+#include <iostream>
+#include <string.h>
+
+void sig_handler(int sig = SIGTERM)
 {
     // Called when exiting or when a signal is received.
-    disable_clock();
-    unSetupDMA();
-    deallocMemPool();
-    unlink(LOCAL_DEVICE_FILE_NAME);
-}
-
-void cleanupAndExit(int sig)
-{
-    // Called when a signal is received. Automatically calls cleanup().
-    prtStdErr("Exiting, caught signal: ", sig, "\n");
-    cleanup();
+    std::cout << "Caught signal: " << sig << " (" << strsignal(sig) << ").\n";
+    dma_cleanup();
     exit(-1);
 }
 
-void setSchedPriority(int priority)
+int main()
 {
-    // In order to get the best timing at a decent queue size, we want the kernel
-    // to avoid interrupting us for long durations.  This is done by giving our
-    // process a high priority. Note, must run as super-user for this to work.
-    struct sched_param sp;
-    sp.sched_priority = priority;
-    int ret = pthread_setschedparam(pthread_self(), SCHED_FIFO, &sp);
-    if (ret)
-    {
-        prtStdErr("Warning: pthread_setschedparam (increase thread priority) returned non-zero: ", ret, "\n");
-    }
-}
-
-void setup_peri_base_virt(volatile unsigned *&peri_base_virt)
-{
-    // Create the memory map between virtual memory and the peripheral range
-    // of physical memory.
-
-    int mem_fd;
-    unsigned gpio_base = gpioBase();
-    // open /dev/mem
-    if ((mem_fd = open("/dev/mem", O_RDWR | O_SYNC)) < 0)
-    {
-        prtStdErr("Error: Can't open /dev/mem\n");
-        exit(-1);
-    }
-    peri_base_virt = (unsigned *)mmap(
-        NULL,
-        0x01000000, // len
-        PROT_READ | PROT_WRITE,
-        MAP_SHARED,
-        mem_fd,
-        gpio_base // base
-    );
-    if ((long int)peri_base_virt == -1)
-    {
-        prtStdErr("Error: peri_base_virt mmap error.\n");
-        exit(-1);
-    }
-    close(mem_fd);
-}
-
-int main(const int argc, char *const argv[])
-{
-    getPLLD(); // Get PLLD Frequency
-
-    update_ppm();
-
-    // Catch all signals (like ctrl+c, ctrl+z, ...) to ensure DMA is disabled
+    // Catch all signals
     for (int i = 0; i < 64; i++)
     {
         struct sigaction sa;
         memset(&sa, 0, sizeof(sa));
-        sa.sa_handler = cleanupAndExit;
+        sa.sa_handler = sig_handler;
         sigaction(i, &sa, NULL);
     }
-    atexit(cleanup);
-    setSchedPriority(30);
 
-    // Initialize the RNG
-    srand(time(NULL));
+    atexit(dma_cleanup);
 
-    // TODO: Push the single hardcoded band to center_freq_set
-    double temp_center_freq_desired;
-    try
-    {
-        temp_center_freq_desired = std::stod(config.frequency_string);
-        config.center_freq_set.push_back(temp_center_freq_desired);
-        if (!std::isfinite(temp_center_freq_desired))
-        {
-            throw std::runtime_error("Invalid floating-point value");
-        }
-    }
-    catch (const std::exception &e)
-    {
-        prtStdErr("Error: Invalid frequency string. Using default 7040100 Hz.\n");
-        config.center_freq_set.push_back(7040100.0);
-    }
-
-    int nbands = config.center_freq_set.size();
-
-    // Initial configuration
-    struct PageInfo constPage;
-    struct PageInfo instrPage;
-    struct PageInfo instrs[1024];
-    setup_peri_base_virt(peri_base_virt);
-
-    // Set up DMA
-    open_mbox();
-    txon();
-    setupDMA(constPage, instrPage, instrs);
-    txoff();
-
-    if (config.mode == TONE)
-    {
-        // Test tone mode...
-        double wspr_symtime = WSPR_SYMTIME;
-        double tone_spacing = 1.0 / wspr_symtime;
-
-        std::stringstream temp;
-        temp << std::setprecision(6) << std::fixed << "Transmitting test tone on frequency " << config.test_tone / 1.0e6 << " MHz.";
-        prtStdOut(temp.str(), "\n");
-        prtStdOut("Press CTRL-C to exit.\n");
-
-        txon(config.useled);
-        int bufPtr = 0;
-        std::vector<double> dma_table_freq;
-        // Set to non-zero value to ensure setupDMATab is called at least once.
-        double ppm_prev = 123456;
-        double center_freq_actual;
-        while (true)
-        {
-            if (config.self_cal)
-            {
-                update_ppm();
-            }
-            if (config.ppm != ppm_prev)
-            {
-                if (config.f_plld_clk <= 0)
-                {
-                    prtStdErr("Error: Invalid PLL clock frequency. Using default 500MHz.\n");
-                    config.f_plld_clk = 500000000.0;
-                }
-                setupDMATab(config.test_tone + 1.5 * tone_spacing, tone_spacing, config.f_plld_clk * (1 - config.ppm / 1e6), dma_table_freq, center_freq_actual, constPage);
-                // cout << std::setprecision(30) << dma_table_freq[0] << "\n";
-                // cout << std::setprecision(30) << dma_table_freq[1] << "\n";
-                // cout << std::setprecision(30) << dma_table_freq[2] << "\n";
-                // cout << std::setprecision(30) << dma_table_freq[3] << "\n";
-                if (center_freq_actual != config.test_tone + 1.5 * tone_spacing)
-                {
-                    std::stringstream temp;
-                    temp << std::setprecision(6) << std::fixed << "Warning: Test tone will be transmitted on " << (center_freq_actual - 1.5 * tone_spacing) / 1e6 << " MHz due to hardware limitations.\n";
-                    prtStdErr(temp.str());
-                }
-                ppm_prev = config.ppm;
-            }
-            txSym(0, center_freq_actual, tone_spacing, 60, dma_table_freq, F_PWM_CLK_INIT, instrs, constPage, bufPtr);
-        }
-    }
-    else
-    {
-        // WSPR mode
-        for (;;)
-        { // Reload Loop >
-            // Create WSPR symbols
-            unsigned char symbols[162];
-            wspr(config.callsign.c_str(), config.locator.c_str(), config.tx_power.c_str(), symbols);
-
-            // // Print encodeed packet
-            // printf("WSPR codeblock: ");
-            // for (int i = 0; i < (signed)(sizeof(symbols)/sizeof(*symbols)); i++) {
-            // if (i) {
-            //     std::cout << ",";
-            // }
-            // printf("%d", symbols[i]);
-            // }
-            // printf("\n");
-
-            prtStdOut("Ready to transmit (setup complete).\n");
-            int band = 0;
-            int n_tx = 0;
-            for (;;)
-            {
-                // Calculate WSPR parameters for this transmission
-                double center_freq_desired;
-                if (config.center_freq_set.empty())
-                {
-                    prtStdErr("Error: center_freq_set is empty. Cannot access band=", band, "\n");
-                    exit(1);
-                }
-                if (band < 0 || band >= static_cast<int>(config.center_freq_set.size()))
-                {
-                    prtStdErr("Error: Band index out of range. band=", band, " size=", config.center_freq_set.size(), "\n");
-                    exit(1);
-                }
-                center_freq_desired = config.center_freq_set[band];
-                bool wspr15 =
-                    (center_freq_desired > 137600 && center_freq_desired < 137625) ||
-                    (center_freq_desired > 475800 && center_freq_desired < 475825) ||
-                    (center_freq_desired > 1838200 && center_freq_desired < 1838225);
-                double wspr_symtime = (wspr15) ? 8.0 * WSPR_SYMTIME : WSPR_SYMTIME;
-                double tone_spacing = 1.0 / wspr_symtime;
-
-                // Add random offset
-                if ((center_freq_desired != 0) && config.random_offset)
-                {
-                    center_freq_desired += (2.0 * rand() / ((double)RAND_MAX + 1.0) - 1.0) * (wspr15 ? WSPR15_RAND_OFFSET : WSPR_RAND_OFFSET);
-                }
-
-                // Status message before transmission
-                std::stringstream temp;
-                temp << std::setprecision(6) << std::fixed;
-                temp << "Desired center frequency for " << (wspr15 ? "WSPR-15" : "WSPR") << " transmission: " << center_freq_desired / 1e6 << " MHz.";
-                prtStdOut(temp.str(), "\n");
-
-                // Wait for WSPR transmission window to arrive.
-                if (config.no_delay)
-                {
-                    prtStdOut("Transmitting immediately (not waiting for WSPR window.)\n");
-                }
-                else
-                {
-                    prtStdOut("Waiting for next WSPR transmission window.\n");
-                    if (!wait_every((wspr15) ? 15 : 2))
-                    {
-                        // Break and reload if ini changes
-                        break;
-                    };
-                }
-
-                // Update crystal calibration information
-                if (config.self_cal)
-                {
-                    update_ppm();
-                }
-
-                // Create the DMA table for this center frequency
-                std::vector<double> dma_table_freq(1024, 0.0); // Ensure vector is properly sized
-                double center_freq_actual;
-                if (center_freq_desired)
-                {
-                    if (config.f_plld_clk <= 0)
-                    {
-                        prtStdErr("Error: Invalid PLL clock frequency. Using default 500MHz.\n");
-                        config.f_plld_clk = 500000000.0;
-                    }
-                    setupDMATab(center_freq_desired, tone_spacing, config.f_plld_clk * (1 - config.ppm / 1e6), dma_table_freq, center_freq_actual, constPage);
-                }
-                else
-                {
-                    center_freq_actual = center_freq_desired;
-                }
-
-                // Send the message if freq != 0 and transmission is enabled
-                if (center_freq_actual && config.xmit_enabled)
-                {
-                    // Print a status message right before transmission begins.
-                    struct timeval tvBegin, tvEnd, tvDiff;
-                    gettimeofday(&tvBegin, NULL);
-                    std::cout << "TX started at: ";
-                    timeval_print(&tvBegin);
-                    std::cout << "\n";
-
-                    struct timeval sym_start;
-                    struct timeval diff;
-                    int bufPtr = 0;
-                    txon(config.useled);
-                    for (int i = 0; i < 162; i++)
-                    {
-                        gettimeofday(&sym_start, NULL);
-                        timeval_subtract(&diff, &sym_start, &tvBegin);
-                        double elapsed = diff.tv_sec + diff.tv_usec / 1e6;
-                        double sched_end = (i + 1) * wspr_symtime;
-                        double this_sym = sched_end - elapsed;
-                        this_sym = (this_sym < .2) ? .2 : this_sym;
-                        this_sym = (this_sym > 2 * wspr_symtime) ? 2 * wspr_symtime : this_sym;
-                        txSym(symbols[i], center_freq_actual, tone_spacing, sched_end - elapsed, dma_table_freq, F_PWM_CLK_INIT, instrs, constPage, bufPtr);
-                    }
-                    n_tx++;
-
-                    // Turn transmitter off
-                    txoff(config.useled);
-
-                    // Time Stamp
-                    gettimeofday(&tvEnd, NULL);
-                    std::cout << "TX ended at: ";
-                    timeval_print(&tvEnd);
-                    timeval_subtract(&tvDiff, &tvEnd, &tvBegin);
-                    printf(" (%ld.%03ld s)\n", tvDiff.tv_sec, (tvDiff.tv_usec + 500) / 1000);
-                }
-                else
-                {
-                    prtStdOut("Skipping transmission.\n");
-                    usleep(1000000);
-                }
-
-                // Advance to next band
-                band = (band + 1) % nbands;
-                if ((band == 0) && !config.repeat)
-                {
-                    return 0;
-                }
-                if ((config.terminate > 0) && (n_tx >= config.terminate))
-                {
-                    return 0;
-                }
-            }
-        } // < Reload Loop
-    }
+    setup_dma();
+    tx_tone();
 
     return 0;
 }
